refactor(식당section): app에서 식당가공해서 section props로 넘겨주도록 변경
